+++ keanu-project/src/test/java/io/improbable/keanu/algorithms/mcmc/testcases/MCMCTestDistributions.java
@@ -26,12 +26,12 @@ public class MCMCTestDistributions {
         DoubleTensor summed = samples.stream()
             .reduce(DoubleTensor.zeros(shape), DoubleTensor::plusInPlace);
 
-        DoubleTensor averages = summed.divInPlace(samples.size());
+        DoubleTensor averages = summed.divInPlace((double) samples.size());
 
         DoubleTensor sumDiffSquared = samples.stream()
             .reduce(
                 DoubleTensor.zeros(shape),
-                (acc, tensor) -> acc.plusInPlace(tensor.minus(averages).powInPlace(2))
+                (acc, tensor) -> acc.plusInPlace(tensor.minus(averages).powInPlace(2.0))
             );
 
         double[] standardDeviations = sumDiffSquared.div(samples.size() - 1).pow(0.5).asFlatDoubleArray();
